ims_auth: fixed compilation and moved back auth_vector to authorize.h
