+++ src/modules/ims_auth/auth_vector.h
@@ -25,33 +25,7 @@
 
 #include "../../core/str.h"
 
-/** Enumeration for the Authorization Vector status */
-enum auth_vector_status
-{
-	AUTH_VECTOR_UNUSED = 0,
-	AUTH_VECTOR_SENT = 1,
-	AUTH_VECTOR_USELESS = 2, /**< invalidated, marked for deletion 		*/
-	AUTH_VECTOR_USED = 3	 /**< the vector has been successfully used	*/
-};
-
-/** Authorization Vector storage structure */
-typedef struct _auth_vector
-{
-	int item_number;	/**< index of the auth vector		*/
-	unsigned char type; /**< type of authentication vector 	*/
-	str authenticate;	/**< challenge (rand|autn in AKA)	*/
-	str authorization;	/**< expected response				*/
-	str ck;				/**< Cypher Key						*/
-	str ik;				/**< Integrity Key					*/
-	time_t expires;		/**< expires in (after it is sent)	*/
-	uint32_t use_nb;	/**< number of use (nonce count) */
-
-	int is_locally_generated; /**< flag to indicate if the vector was generated locally, not on the HSS */
-
-	enum auth_vector_status status; /**< current status		*/
-	struct _auth_vector *next;		/**< next av in the list		*/
-	struct _auth_vector *prev;		/**< previous av in the list	*/
-} auth_vector;
+#include "authorize.h"
 
 auth_vector *auth_vector_make_local(uint8_t k[16], uint8_t op[16], int opIsOPc,
 		uint8_t amf[2], uint8_t sqn[6]);
