+++ src/modules/ims_auth/cxdx_mar.h
@@ -47,6 +47,8 @@
 
 #include "../../core/mod_fix.h"
 
+#include "../cdp/diameter.h"
+
 struct sip_msg;
 struct parsed_maa;
